@alejandropg 2020-06-18
I supposed it. In the entire world we are not in our best moment 😓
Really thanks a lot for your time and sorry for my "ping".